backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.4.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
